requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Session==0.5.0
APScheduler==3.11.1
requests==2.31.0
urllib3==2.1.0
cryptography>=46.0.0
bcrypt==4.1.2
dnspython==2.4.2
gunicorn==21.2.0

# TimescaleDB / PostgreSQL support (v2.0.0)
psycopg2-binary==2.9.9
psycopg2-pool==1.1
